+++ src/utils/VirtruWrapper.js
@@ -174,6 +174,7 @@ function newVirtruDecryptParamsBuilder(opts) {
 
 async function signOut(userId) {
   const resetApp = () => {
+    localStorage.removeItem('virtru-demo-email');
     localStorage.removeItem('virtru-demo-file');
     localStorage.removeItem('virtru-demo-policy');
     localStorage.removeItem('virtru-demo-sdk-log');